worker/requirements.txt
python-dotenv
paho.mqtt
pyserial
influxdb
